update subscription price yupe

+++ backend/common/ent/schema/subscription.go
@@ -18,7 +18,7 @@ type Subscription struct {
 func (Subscription) Fields() []ent.Field {
 	return []ent.Field{
 		field.String("provider_plan_id"),
-		field.Int("price"),
+		field.Float("price"),
 		field.Int("duration_in_months"),
 		field.Bool("is_active"),
 		field.String("name"),
